+++ src/main/java/org/opentripplanner/model/impl/OtpTransitServiceImpl.java
@@ -123,13 +123,6 @@ class OtpTransitServiceImpl implements OtpTransitService {
     this.tripPatterns = immutableList(builder.getTripPatterns().values());
     this.trips = immutableList(builder.getTripsById().values());
     this.flexTrips = immutableList(builder.getFlexTripsById().values());
-
-    if (!builder.getFrequencies().isEmpty()) {
-      LOG.error(
-        "OTP2 do not support GTFS Trip Frequencies. " +
-        "See https://github.com/opentripplanner/OpenTripPlanner/issues/3243."
-      );
-    }
   }
 
   @Override