+++ src/main/java/org/opentripplanner/routing/algorithm/raptoradapter/router/street/DirectStreetRouter.java
@@ -49,7 +49,7 @@ public class DirectStreetRouter {
         router.graph.ellipsoidToGeoidDifference
       );
       List<Itinerary> response = graphPathToItineraryMapper.mapItineraries(paths);
-      ItinerariesHelper.decorateItinerariesWithRequestData(response, routingContext);
+      ItinerariesHelper.decorateItinerariesWithRequestData(response, request);
       return response;
     } catch (PathNotFoundException e) {
       return Collections.emptyList();