+++ docs/Changelog.md
@@ -37,6 +37,9 @@ based on merged pull requests. Search GitHub issues and pull requests for smalle
 - Add an implementation for fetching alternative legs [#4071](https://github.com/opentripplanner/OpenTripPlanner/pull/4071)
 - Group/filter on first/last trip in itineraries [#4051](https://github.com/opentripplanner/OpenTripPlanner/pull/4051)
 - Allow single SIRI-situations to fail without affecting others [#4083](https://github.com/opentripplanner/OpenTripPlanner/pull/4083)
+- Read max slope from default routing request [#4098](https://github.com/opentripplanner/OpenTripPlanner/pull/4098)
+- Fix problem with submode filter [#4093](https://github.com/opentripplanner/OpenTripPlanner/pull/4093)
+- Make maxWheelchairSlope not a hard cut-off, but apply a cost instead [#4088](https://github.com/opentripplanner/OpenTripPlanner/pull/4088)
   [](AUTOMATIC_CHANGELOG_PLACEHOLDER_DO_NOT_REMOVE)
 
 ## 2.1.0 (2022-03-17)